# Improve log injection perf for NLog 4.6+