--- a/.azure-pipelines/integration-tests.yml
+++ b/.azure-pipelines/integration-tests.yml
@@ -223,6 +223,7 @@ jobs:
     inputs:
       command: build
       projects: |
+        benchmarks/**/*.csproj
         customer-samples/**/*.csproj
         reproductions/**/*.csproj
         samples/**/*.csproj