--- a/benchmarks/Directory.Build.props
+++ b/benchmarks/Directory.Build.props
@@ -1,0 +1,6 @@
+<Project>
+  <!--
+  This file intentionally left blank...
+  to stop msbuild from looking up the folder hierarchy
+  -->
+</Project>