--- a/benchmarks/NLog46/Benchmarks.cs
+++ b/benchmarks/NLog46/Benchmarks.cs
@@ -40,7 +40,6 @@ namespace NLog46
             var newLogInjectionSettings = new TracerSettings
             {
                 LogsInjectionEnabled = true,
-                UseWebServerResourceAsOperationName = false,
             };
             NewLogInjectionTracer = new Tracer(newLogInjectionSettings);