--- a/benchmarks/NLog46/Main.cs
+++ b/benchmarks/NLog46/Main.cs
@@ -1,0 +1,9 @@
+using BenchmarkDotNet.Running;
+
+namespace NLog46
+{
+    class Program
+    {
+        static void Main(string[] args) => BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
+    }
+}