--- a/customer-samples/AutomaticTraceIdInjection/NLog46Example/NLog46Example.csproj
+++ b/customer-samples/AutomaticTraceIdInjection/NLog46Example/NLog46Example.csproj
@@ -2,11 +2,11 @@
    <!-- Modified by SignalFx --> 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="NLog" Version="4.6.7" />
+    <PackageReference Include="NLog" Version="4.6.8" />
   </ItemGroup>
 
   <ItemGroup>
@@ -17,6 +17,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\src\Datadog.Trace\Datadog.Trace.csproj" />
+    <ProjectReference Include="..\..\..\src\Datadog.Trace.OpenTracing\Datadog.Trace.OpenTracing.csproj" />
   </ItemGroup>
 
 </Project>