--- a/customer-samples/AutomaticTraceIdInjection/NLog46Example/Program.cs
+++ b/customer-samples/AutomaticTraceIdInjection/NLog46Example/Program.cs
@@ -1,5 +1,6 @@
-using Datadog.Trace;
+// Modified by SignalFx
 using NLog;
+using SignalFx.Tracing;
 
 namespace NLog46Example
 {