--- a/src/Datadog.Trace/Logging/LibLogScopeEventSubscriber.cs
+++ b/src/Datadog.Trace/Logging/LibLogScopeEventSubscriber.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Concurrent;
 using System.Diagnostics;
+using SignalFx.Tracing.Configuration;
 using SignalFx.Tracing.Logging.LogProviders;
 
 namespace SignalFx.Tracing.Logging
@@ -40,23 +41,54 @@ namespace SignalFx.Tracing.Logging
         //            but the target AppDomain is unable to de-serialize the object --
         //            this can easily happen if the target AppDomain cannot find/load the
         //            logging framework assemblies.
-        public LibLogScopeEventSubscriber(IScopeManager scopeManager)
+        public LibLogScopeEventSubscriber(IScopeManager scopeManager, TracerSettings settings)
         {
             _scopeManager = scopeManager;
 
             _logProvider = LogProvider.CurrentLogProvider ?? LogProvider.ResolveLogProvider();
-            if (_logProvider is SerilogLogProvider)
+            switch (_logProvider)
             {
-                // Do not set default values for Serilog because it is unsafe to set
-                // except at the application startup, but this would require auto-instrumentation
-                _scopeManager.SpanOpened += StackOnSpanOpened;
-                _scopeManager.SpanClosed += StackOnSpanClosed;
+                case SerilogLogProvider:
+                    // Do not set default values for Serilog because it is unsafe to set
+                    // except at the application startup, but this would require auto-instrumentation
+                    _scopeManager.SpanOpened += StackOnSpanOpened;
+                    _scopeManager.SpanClosed += StackOnSpanClosed;
+                    break;
+
+                case NLogLogProvider when UseNLogOptimized():
+                    UseValueReaders(settings);
+                    break;
+
+                default:
+                    _scopeManager.SpanActivated += MapOnSpanActivated;
+                    _scopeManager.TraceEnded += MapOnTraceEnded;
+                    break;
             }
-            else
+        }
+
+        public static bool UseNLogOptimized()
+        {
+            // This code checks the same requisits from LibLog\5.0.6\LogProviders\NLogLogProvider.cs to see
+            // if the code can use an optimized path for NLog.
+            var ndlcContextType = Type.GetType("NLog.NestedDiagnosticsLogicalContext, NLog");
+            if (ndlcContextType != null)
             {
-                _scopeManager.SpanActivated += MapOnSpanActivated;
-                _scopeManager.TraceEnded += MapOnTraceEnded;
+                var pushObjectMethod = ndlcContextType.GetMethod("PushObject", typeof(object));
+                if (pushObjectMethod != null)
+                {
+                    var mdlcContextType = Type.GetType("NLog.MappedDiagnosticsLogicalContext, NLog");
+                    if (mdlcContextType != null)
+                    {
+                        var setScopedMethod = mdlcContextType.GetMethod("SetScoped", typeof(string), typeof(object));
+                        if (setScopedMethod != null)
+                        {
+                            return true;
+                        }
+                    }
+                }
             }
+
+            return false;
         }
 
         public void StackOnSpanOpened(object sender, SpanEventArgs spanEventArgs)
@@ -95,6 +127,39 @@ namespace SignalFx.Tracing.Logging
             }
 
             RemoveAllCorrelationIdentifierContexts();
+        }
+
+        private void UseValueReaders(TracerSettings settings)
+        {
+            var traceIdReader = new ValueReader(() =>
+            {
+                return _scopeManager.Active?.Span?.TraceId.ToString() ?? TraceId.Zero.ToString();
+            });
+            LogProvider.OpenMappedContext(
+                CorrelationIdentifier.TraceIdKey, traceIdReader, destructure: false);
+
+            var spanIdReader = new ValueReader(() =>
+            {
+                return _scopeManager.Active?.Span?.SpanId.ToString("x16") ?? "0000000000000000";
+            });
+            LogProvider.OpenMappedContext(
+                CorrelationIdentifier.SpanIdKey, spanIdReader, destructure: false);
+
+            var serviceReader = new ValueReader(() =>
+            {
+                // It is possible to have service name per span so try that first, anyway it should not
+                // fail but use settings as a fallback just in case.
+                return _scopeManager.Active?.Span?.ServiceName ?? settings.ServiceName;
+            });
+            LogProvider.OpenMappedContext(
+                CorrelationIdentifier.ServiceNameKey, serviceReader, destructure: false);
+
+            var environmentReader = new ValueReader(() =>
+            {
+                return settings.Environment ?? string.Empty;
+            });
+            LogProvider.OpenMappedContext(
+                CorrelationIdentifier.ServiceEnvironmentKey, environmentReader, destructure: false);
         }
 
         private void RemoveLastCorrelationIdentifierContext()
@@ -152,6 +217,21 @@ namespace SignalFx.Tracing.Logging
             {
                 _safeToAddToMdc = false;
                 RemoveAllCorrelationIdentifierContexts();
+            }
+        }
+
+        private class ValueReader
+        {
+            private readonly Func<string> readerFunc;
+
+            public ValueReader(Func<string> reader)
+            {
+                readerFunc = reader;
+            }
+
+            public override string ToString()
+            {
+                return readerFunc();
             }
         }