--- a/test/Datadog.Trace.Tests/Sampling/RateLimiterTests.cs
+++ b/test/Datadog.Trace.Tests/Sampling/RateLimiterTests.cs
@@ -50,7 +50,7 @@ namespace Datadog.Trace.Tests.Sampling
             Assert.Equal(expected: DefaultLimitPerSecond, actual: allowedCount);
         }
 
-        [Fact]
+        [Fact(Skip = "Flaky test")]
         public void Limits_Approximately_To_Defaults()
         {
             Run_Limit_Test(intervalLimit: null, numberPerBurst: 200, numberOfBursts: 20, millisecondsBetweenBursts: 247);