A very targeted optimization for NLog 4.6+ we probably can do similar for other loggers but the scope of this one is targeted to this specific version of NLog. The optimization is very simple: instead of creating and disposing of logging context objects the new code leverages the active scope, so there is no churn of objects to update the context.

BenchmarkDotNet=v0.12.1, OS=ubuntu 20.04
Intel Core i7-8665U CPU 1.90GHz (Coffee Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.203
  [Host]     : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT
  DefaultJob : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT


|                     Method |         Categories |     Mean |    Error |   StdDev |   Median | Ratio | RatioSD |  Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|--------------------------- |------------------- |---------:|---------:|---------:|---------:|------:|--------:|-------:|-------:|-------:|----------:|
|                  NoTagging | NoActiveSpanSwitch | 11.67 us | 0.339 us | 0.944 us | 11.37 us |  1.00 |    0.00 | 0.5951 | 0.2747 | 0.0458 |   3.53 KB |
|                 OldTagging | NoActiveSpanSwitch | 23.30 us | 1.204 us | 3.532 us | 21.91 us |  2.01 |    0.31 | 1.5869 | 0.4883 | 0.1221 |   9.06 KB |
|           NewNLog46Tagging | NoActiveSpanSwitch | 12.57 us | 0.450 us | 1.291 us | 12.36 us |  1.08 |    0.15 | 0.5951 | 0.3052 | 0.0305 |   3.59 KB |
|                            |                    |          |          |          |          |       |         |        |        |        |           |
|        NoTaggingSpanSwitch |   ActiveSpanSwitch | 18.93 us | 1.034 us | 3.034 us | 18.03 us |  1.00 |    0.00 | 0.7935 | 0.3967 | 0.0916 |   4.78 KB |
|       OldTaggingSpanSwitch |   ActiveSpanSwitch | 54.23 us | 2.928 us | 8.448 us | 53.87 us |  2.91 |    0.60 | 2.5635 | 0.3662 |      - |  15.88 KB |
| NewNLog46TaggingSpanSwitch |   ActiveSpanSwitch | 17.29 us | 0.909 us | 2.594 us | 16.61 us |  0.93 |    0.21 | 0.8545 | 0.4272 | 0.0610 |   4.84 KB |


BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-8665U CPU 1.90GHz (Coffee Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.203
  [Host]     : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT
  DefaultJob : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT


|                     Method |         Categories |      Mean |     Error |    StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|--------------------------- |------------------- |----------:|----------:|----------:|------:|--------:|-------:|-------:|-------:|----------:|
|                  NoTagging | NoActiveSpanSwitch |  9.066 us | 0.3204 us | 0.9192 us |  1.00 |    0.00 | 0.6104 | 0.3052 | 0.2136 |   3.53 KB |
|                 OldTagging | NoActiveSpanSwitch | 15.865 us | 0.4551 us | 1.3348 us |  1.77 |    0.25 | 1.5259 | 0.3967 | 0.1526 |   9.06 KB |
|           NewNLog46Tagging | NoActiveSpanSwitch |  8.160 us | 0.2620 us | 0.7559 us |  0.91 |    0.13 | 0.5951 | 0.3052 | 0.1373 |   3.59 KB |
|                            |                    |           |           |           |       |         |        |        |        |           |
|        NoTaggingSpanSwitch |   ActiveSpanSwitch | 12.060 us | 0.3307 us | 0.9700 us |  1.00 |    0.00 | 0.7629 | 0.3662 | 0.1831 |   4.78 KB |
|       OldTaggingSpanSwitch |   ActiveSpanSwitch | 27.664 us | 1.3919 us | 4.0382 us |  2.30 |    0.34 | 2.7466 | 0.5493 | 0.1831 |  15.88 KB |
| NewNLog46TaggingSpanSwitch |   ActiveSpanSwitch | 12.056 us | 0.2931 us | 0.8362 us |  1.00 |    0.10 | 0.7935 | 0.3967 | 0.1831 |   4.84 KB |